+++ api/config/crd/bases/odigos.io_destinations.yaml
@@ -57,11 +57,6 @@ spec:
                   type: string
                 type: array
               type:
-                enum:
-                - grafana
-                - datadog
-                - honeycomb
-                - newrelic
                 type: string
             required:
             - data
