+++ autoscaler/controllers/gateway/config/grafana.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	odigosv1 "github.com/keyval-dev/odigos/api/v1alpha1"
 	commonconf "github.com/keyval-dev/odigos/autoscaler/controllers/common"
+	"github.com/keyval-dev/odigos/common"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 	"strings"
 )
@@ -14,8 +15,8 @@ const (
 
 type Grafana struct{}
 
-func (g *Grafana) DestType() odigosv1.DestinationType {
-	return odigosv1.GrafanaDestinationType
+func (g *Grafana) DestType() common.DestinationType {
+	return common.GrafanaDestinationType
 }
 
 func (g *Grafana) ModifyConfig(dest *odigosv1.Destination, currentConfig *commonconf.Config) {
