+++ common/dests.go
@@ -1,0 +1,10 @@
+package common
+
+type DestinationType string
+
+const (
+	GrafanaDestinationType   DestinationType = "grafana"
+	DatadogDestinationType   DestinationType = "datadog"
+	HoneycombDestinationType DestinationType = "honeycomb"
+	NewRelicDestinationType  DestinationType = "newrelic"
+)
